Refactor worker structure
