+++ cmd/healthcheck/main.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/golang/glog"
 	"github.com/kingcobra2468/cot/internal/config"
-	"github.com/kingcobra2468/cot/internal/router"
-	"github.com/kingcobra2468/cot/internal/router/gvoice"
+	"github.com/kingcobra2468/cot/internal/router/worker"
+	"github.com/kingcobra2468/cot/internal/router/worker/gvoice"
 	"github.com/spf13/viper"
 )
 
@@ -39,7 +39,7 @@ func parseGVMS() (*config.GVMS, error) {
 
 // ping checks if COT is online.
 func ping(number string) error {
-	l := router.NewTextWorker(gvoice.Link{GVoiceNumber: "14159422253", ClientNumber: "14159422253"}, false, 0)
+	l := worker.NewGVoiceWorker(gvoice.Link{GVoiceNumber: "14159422253", ClientNumber: "14159422253"}, false, 0)
 	if err := l.Send("ping"); err != nil {
 		glog.Error(err)
 		return nil
